Update instructions on installing Pillow-SIMD
It's best to install gxx_linux-64 first, before libjpeg-turbo as this pulls in a bunch of deps. We then uninstall jpeg libtiff and pillow which causes the environment to become inconsistent, then we have to be very careful. We force install of libjpeg-turbo using `--no-deps` (without this jpeg and libtiff would be brought back in), then we uninstall pillow via pip so as to trick conda into thinking its still installed. Then we replace pillow using our compiled version and then restore jpeg and libtiff to bring the environment back into consistency