Speed up dropping of problematic metadata fields
There was a loop with
```
self.dataset.video_records[i]
```
inside, which accesses the `video_records` property on the dataset.
This property recreates the entire list of video records on each call,
so is very expensive. I have avoided doing this, and have also changed
the behaviour so that this call accesses an internal property of the
list of video records which is set in the constructor so that this
performance hit won't occur again.